Install cert-manager for e2e tests
Some eventing e2e tests are using eventhub.WithTLS(), which creates
certificates and thus requires cert-manager in their setup. Even this
tests might be skipped, as this is run in their environment setup,
theses test fail if cert-manager is not installed. Therefor we need to
install cert-manager for the e2e tests.