Update dependencies to latest (#29)
* Update dependencies to latest

* Remove dependency to knative.dev/pkg/client/injection/kube/client/client_expansion.go

Codegen support for informers and clients that use the dynamic.Client interface
was dropped in https://github.com/knative/pkg/pull/2742.

* Run update-codegen.sh

* Regenerate pkg/client/injection/kube

* Update 3rd party eventing-latest

* Install cert-manager for e2e tests

Some eventing e2e tests are using eventhub.WithTLS(), which creates
certificates and thus requires cert-manager in their setup. Even this
tests might be skipped, as this is run in their environment setup,
theses test fail if cert-manager is not installed. Therefor we need to
install cert-manager for the e2e tests.

* Update eventing core dependency

* Fix shell lint issue